requirements.txt
asgiref==3.8.1
attrs==25.3.0
cfgv==3.4.0
distlib==0.3.9
Django==5.1.7
django-autoslug==1.9.9
django-shell-plus==1.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
filelock==3.18.0
identify==2.6.9
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.1.0
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.29.3
